+++ app/keepers/keepers.go
@@ -385,6 +385,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BankKeeper,
 		ak.AccountKeeper,
 		&epochingKeeper,
+		ak.FeeGrantKeeper,
 		appparams.AccGov.String(),
 		authtypes.FeeCollectorName,
 	)