+++ testutil/keeper/incentive.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"testing"
+	"time"
 
 	"cosmossdk.io/core/header"
 	"cosmossdk.io/log"
@@ -30,6 +31,7 @@ func IncentiveKeeperWithStore(
 	bankKeeper types.BankKeeper,
 	accountKeeper types.AccountKeeper,
 	epochingKeeper types.EpochingKeeper,
+	feegrantKeeper types.FeegrantKeeper,
 ) (*keeper.Keeper, sdk.Context) {
 	if storeKey == nil {
 		storeKey = storetypes.NewKVStoreKey(types.StoreKey)
@@ -47,18 +49,26 @@ func IncentiveKeeperWithStore(
 		bankKeeper,
 		accountKeeper,
 		epochingKeeper,
+		feegrantKeeper,
 		appparams.AccGov.String(),
 		authtypes.FeeCollectorName,
 	)
 
-	ctx := sdk.NewContext(stateStore, cmtproto.Header{}, false, log.NewNopLogger())
+	ctx := sdk.NewContext(
+		stateStore,
+		cmtproto.Header{
+			Time: time.Now().UTC(),
+		},
+		false,
+		log.NewNopLogger(),
+	)
 	ctx = ctx.WithHeaderInfo(header.Info{})
 
 	return &k, ctx
 }
 
-func IncentiveKeeper(t testing.TB, bankKeeper types.BankKeeper, accountKeeper types.AccountKeeper, epochingKeeper types.EpochingKeeper) (*keeper.Keeper, sdk.Context) {
-	return IncentiveKeeperWithStoreKey(t, nil, bankKeeper, accountKeeper, epochingKeeper)
+func IncentiveKeeper(t testing.TB, bankKeeper types.BankKeeper, accountKeeper types.AccountKeeper, epochingKeeper types.EpochingKeeper, feegrantKeeper types.FeegrantKeeper) (*keeper.Keeper, sdk.Context) {
+	return IncentiveKeeperWithStoreKey(t, nil, bankKeeper, accountKeeper, epochingKeeper, feegrantKeeper)
 }
 
 func IncentiveKeeperWithStoreKey(
@@ -67,11 +77,12 @@ func IncentiveKeeperWithStoreKey(
 	bankKeeper types.BankKeeper,
 	accountKeeper types.AccountKeeper,
 	epochingKeeper types.EpochingKeeper,
+	feegrantKeeper types.FeegrantKeeper,
 ) (*keeper.Keeper, sdk.Context) {
 	db := dbm.NewMemDB()
 	stateStore := store.NewCommitMultiStore(db, log.NewTestLogger(t), storemetrics.NewNoOpMetrics())
 
-	k, ctx := IncentiveKeeperWithStore(t, db, stateStore, storeKey, bankKeeper, accountKeeper, epochingKeeper)
+	k, ctx := IncentiveKeeperWithStore(t, db, stateStore, storeKey, bankKeeper, accountKeeper, epochingKeeper, feegrantKeeper)
 
 	// Initialize params
 	if err := k.SetParams(ctx, types.DefaultParams()); err != nil {