+++ x/incentive/keeper/intercept_fee_collector_test.go
@@ -41,7 +41,9 @@ func FuzzInterceptFeeCollector(f *testing.F) {
 		// mock epoching keeper
 		epochingKeeper := types.NewMockEpochingKeeper(ctrl)
 
-		keeper, ctx := testkeeper.IncentiveKeeper(t, bankKeeper, accountKeeper, epochingKeeper)
+		feegrantKeeper := types.NewMockFeegrantKeeper(ctrl)
+
+		keeper, ctx := testkeeper.IncentiveKeeper(t, bankKeeper, accountKeeper, epochingKeeper, feegrantKeeper)
 		height := datagen.RandomInt(r, 1000)
 		ctx = datagen.WithCtxHeight(ctx, height)
 
@@ -84,7 +86,9 @@ func TestInterceptFeeCollectorWithSmallAmount(t *testing.T) {
 	// mock epoching keeper
 	epochingKeeper := types.NewMockEpochingKeeper(ctrl)
 
-	keeper, ctx := testkeeper.IncentiveKeeper(t, bankKeeper, accountKeeper, epochingKeeper)
+	feegrantKeeper := types.NewMockFeegrantKeeper(ctrl)
+
+	keeper, ctx := testkeeper.IncentiveKeeper(t, bankKeeper, accountKeeper, epochingKeeper, feegrantKeeper)
 	height := datagen.RandomInt(r, 1000)
 	ctx = datagen.WithCtxHeight(ctx, height)