+++ x/incentive/keeper/params_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 func TestGetParams(t *testing.T) {
-	k, ctx := testkeeper.IncentiveKeeper(t, nil, nil, nil)
+	k, ctx := testkeeper.IncentiveKeeper(t, nil, nil, nil, nil)
 	params := types.DefaultParams()
 
 	err := k.SetParams(ctx, params)